% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualenrollmentApp.R
\name{visualenrollmentApp}
\alias{visualenrollmentApp}
\title{Run Shiny app}
\usage{
visualenrollmentApp(...)
}
\arguments{
\item{...}{optional parameters for the shinyApp call}
}
\description{
Run Shiny app
}
